+++ cli/tools/mod.rs
@@ -5,5 +5,6 @@ pub mod fmt;
 pub mod installer;
 pub mod lint;
 pub mod repl;
+pub mod standalone;
 pub mod test_runner;
 pub mod upgrade;
